backend/requirements.txt
celery==4.3.0
Django==2.0
djangorestframework==3.7.3
gunicorn==19.7.1
kafka-python==1.4.7
mysqlclient==1.3.12
redis==3.2.0
django-cors-headers==3.1.1
django-celery-beat==1.5.0